tests/wpt/tests/document-picture-in-picture/enter-event.https.html

<!DOCTYPE html>
<title>Test that enter event is fired at documentPictureInPicture
  when the PiP window in opened.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<body>
<script>
async_test((t) => {
  test_driver.bless('request PiP window').then(t.step_func(_ => {
    documentPictureInPicture.onenter = t.step_func_done((e) => {
      assert_true(e.isTrusted, "is trusted");
      assert_false(e.cancelable, "is not cancelable");
    });
    documentPictureInPicture.requestWindow();
  }));
});
</script>
</body>
